+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 Remember to bring your dependencies up to date with `./scripts/venvinstall.sh` when updating to this version!
 
+- Minor: Moved chat alerts into a sub-category. Also added a new raid alert module
 - Minor: Added option to unban user from chat when the `!unpermaban` command is used (#739, #887)
 - Minor: Added whisper timeout reasons to the emote limit module (#866)
 - Minor: Added `timeout_end` field to User API Response and `!debug user` command output. (#875)